--- src/agent/tools/langchain.pdf.tool.ts
@@ -1,33 +1,0 @@
-// import { Tool } from "langchain/tools";
-// import { PDFLoader } from "langchain/document_loaders/fs/pdf";
-// import { RecursiveCharacterTextSplitter } from "langchain/text_splitter";
-// import { OpenAIEmbeddings } from "langchain/embeddings/openai";
-// import { MemoryVectorStore } from "langchain/vectorstores/memory";
-
-// export class LangChainPDFTool extends Tool {
-//   name = "pdf_processor";
-//   description = "Process PDF files and extract structured content with embeddings";
-
-//   async _call(filePath: string): Promise<string> {
-//     // Load PDF
-//     const loader = new PDFLoader(filePath);
-//     const docs = await loader.load();
-
-//     // Split text into chunks
-//     const textSplitter = new RecursiveCharacterTextSplitter({
-//       chunkSize: 1000,
-//       chunkOverlap: 200,
-//     });
-//     const splitDocs = await textSplitter.splitDocuments(docs);
-
-//     // Create embeddings
-//     const embeddings = new OpenAIEmbeddings();
-//     const vectorStore = await MemoryVectorStore.fromDocuments(splitDocs, embeddings);
-
-//     return JSON.stringify({
-//       chunks: splitDocs.length,
-//       content: splitDocs.map(doc => doc.pageContent),
-//       metadata: splitDocs.map(doc => doc.metadata)
-//     });
-//   }
-// }
