+++ src/agent/upload.agent.ts
@@ -1,5 +1,5 @@
 import { llm } from "@services/ollama.service"
 
-import { ScraperTool } from "./tools/langchain.scraper.tool";
+import { ScraperTool } from "./tools/scraper.tool";
 
 export const agent = llm.bindTools([ScraperTool])
